Joining partners now unions files rather than deleting the second partner's files.

+++ src/model/Partner_request.C
@@ -29,27 +29,47 @@ Partner_request::create(Session& session,
     return session.add(request);
 }
 
-void Partner_request::confirm(Session& session) const
+bool Partner_request::confirm(Session& session) const
 {
+    // Confirm that the request hasn't been cancelled
+    auto self = Partner_request::find_by_requestor_and_requestee(
+            session, requestor_, requestee_, assignment_);
+    if (!self) return false;
+
     auto submission1 = Submission::find_by_assignment_and_user(
             session, assignment_, requestor_);
     auto submission2 = Submission::find_by_assignment_and_user(
             session, assignment_, requestee_);
 
     if (submission1->user2() || submission2->user2())
-        return;
+        return false;
 
-    submission1.modify()->set_user2(requestee_);
-    submission2.remove();
+    bool success = Submission::join_together(submission1, submission2);
 
-    session.execute("DELETE FROM partner_requests"
-                    " WHERE (requestor_id = ? OR requestor_id = ?"
-                    "    OR  requestee_id = ? OR requestee_id = ?)"
-                    "   AND assignment_number = ?")
-           .bind(requestor_.id()).bind(requestee_.id())
-           .bind(requestor_.id()).bind(requestee_.id())
-           .bind(assignment_.id())
-           .run();
+    if (success)
+        session.execute("DELETE FROM partner_requests"
+                        " WHERE (requestor_id = ? OR requestor_id = ?"
+                        "    OR  requestee_id = ? OR requestee_id = ?)"
+                        "   AND assignment_number = ?")
+               .bind(requestor_.id()).bind(requestee_.id())
+               .bind(requestor_.id()).bind(requestee_.id())
+               .bind(assignment_.id())
+               .run();
+
+    return success;
+}
+
+Wt::Dbo::ptr<Partner_request>
+Partner_request::find_by_requestor_and_requestee(
+        Session& session,
+        const Wt::Dbo::ptr<User>& requestor,
+        const Wt::Dbo::ptr<User>& requestee,
+        const Wt::Dbo::ptr<Assignment>& assignment)
+{
+    return session.find<Partner_request>()
+                  .where("requestor_id = ?").bind(requestor.id())
+                  .where("requestee_id = ?").bind(requestee.id())
+                  .where("assignment_number = ?").bind(assignment.id());
 }
 
 Wt::Dbo::ptr<Partner_request>
